Refactoring code and adding test for example
