+++ flash_examples/question_answering.py
@@ -28,8 +28,7 @@ datamodule = QuestionAnsweringData.from_squad_v2(
 model = QuestionAnsweringTask()
 
 # 3. Create the trainer and finetune the model
-# trainer = Trainer(max_epochs=3)
-trainer = Trainer(fast_dev_run=True)
+trainer = Trainer(max_epochs=3)
 trainer.finetune(model, datamodule=datamodule)
 
 # 4. Summarize some text!
